+++ packages/inventor-plugin-app-react-webpack/lib/actions/dev.d.ts
@@ -1,0 +1,22 @@
+/**
+ * action 入口
+ * @author: sunkeysun
+ */
+import { Action } from '@inventorjs/cli-core';
+import webpackDevServer from 'webpack-dev-server';
+type DevServerConfig = webpackDevServer.Configuration;
+interface ServerInfo {
+    localAddress: string;
+    staticPath: DevServerConfig['static'];
+    historyApiFallback: DevServerConfig['historyApiFallback'];
+}
+export default class DevAction extends Action {
+    description: string;
+    options: {
+        flags: string;
+        description: string;
+    }[];
+    logServerInfo({ localAddress, staticPath, historyApiFallback }: ServerInfo): void;
+    run(_: string[], options: Record<string, unknown>): Promise<void>;
+}
+export {};