+++ packages/inventor-plugin-app-react-webpack/lib/index.d.ts
@@ -1,0 +1,8 @@
+/**
+ * 插件入口
+ * @author: sunkeysun
+ */
+import { Plugin } from '@inventorjs/cli-core';
+export default class AppPlugin extends Plugin {
+    description: string;
+}