+++ packages/inventor-plugin-app-react-webpack/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @inventorjs/plugin-app-react-webpack
 
+## 5.1.2
+
+### Patch Changes
+
+- change to 5.1.1
+- Updated dependencies
+  - @inventorjs/cli-core@5.1.2
+
+## 6.0.1
+
+### Patch Changes
+
+- correct version
+- Updated dependencies
+  - @inventorjs/cli-core@5.1.1
+
+## 6.0.0
+
+### Minor Changes
+
+- change dist
+
+### Patch Changes
+
+- Updated dependencies
+  - @inventorjs/cli-core@5.1.0
+
 ## 5.0.18
 
 ### Patch Changes